requirements.txt
# Ingestion
yfinance==0.2.38
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# NLP
transformers==4.40.0
torch==2.2.2
vaderSentiment==3.3.2
deep-translator==1.11.4

# Data processing
pandas==2.2.1
numpy==1.26.4

# Econometrics
statsmodels==0.14.1
scipy==1.13.0

# Visualization
plotly==5.20.0
dash==2.16.1
matplotlib==3.8.4
seaborn==0.13.2

# Utils
python-dotenv==1.0.1
loguru==0.7.2
pyyaml==6.0.1

# Tests
pytest==8.1.1
